requirements.txt
opencv-python
pillow
numpy
tkinter
